Add Collsion Interface
Add 4 interfaces and 1 stand alone class HitData.

Inherit the hurt responder interface in Botcontroller script

Add hitbox scripts for GunPart bullet projectile and SwordPart Hitboxes

Inherit hitresponder interface in Gunpart and Swordpart scripts
