+++ Assets/Prefabs/BotPartPrefabs/New System/NewSystem_HitHurtBox_TEST.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 321c7099883000043b1cd7ff1d53adea
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
