+++ Assets/Prefabs/BotPartPrefabs/New System/NewSystem_HitHurtBox_TEST/DrillPart 1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e3d336448a75ca24db59b74a5184a7c2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
