+++ Assets/Prefabs/BotPartPrefabs/New System/NewSystem_HitHurtBox_TEST/GunParts/Collision_bullet.prefab
@@ -153,7 +153,7 @@ MonoBehaviour:
   m_layerMask:
     serializedVersion: 2
     m_Bits: 2048
-  m_hurtboxMask: 0
+  m_hurtboxMask: 2
 --- !u!114 &-7950421402837390847
 MonoBehaviour:
   m_ObjectHideFlags: 0
