+++ Assets/Prefabs/BotPartPrefabs/New System/NewSystem_HitHurtBox_TEST/SwordPart 1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3b16a5656207cc348a02a42f3c1626ec
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
