+++ Assets/Prefabs/BotPartPrefabs/New System/NewSystem_HitHurtBox_TEST/TailFinPart 1.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 0422905f6572fc54095557250e1fa731
-folderAsset: yes
-DefaultImporter:
+guid: a4e7ac0cf1086f349b6f734f7f1714da
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
