+++ Assets/Prefabs/BotPartPrefabs/New System/NewSystem_HitHurtBox_TEST/TeleporterPart 1.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 3d55f6fb72facf34f8f795265bcdf8b0
-folderAsset: yes
-DefaultImporter:
+guid: 4f9907a2c1ca5d6448ca8b942c81b504
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
