+++ Assets/Prefabs/BotPartPrefabs/New System/NewSystem_HitHurtBox_TEST/TeslaPart 1.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 54fbe270871b8b649a0cd53d51d9915d
-folderAsset: yes
-DefaultImporter:
+guid: fbdfd7e16e40d5a4e987b21cb479dde4
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
