+++ Assets/Prefabs/BotPartPrefabs/New System/NewSystem_HitHurtBox_TEST/WheelPart 1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ddbcf14ba66704941a6849e3da0b6cde
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
