+++ Assets/Prefabs/BotPartPrefabs/New System/NewSystem_HitHurtBox_TEST/WingsPart 1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f1f2f414dc4ec07448e35cfbd7a074e0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
