+++ Assets/Scripts/Bot Functionality/Bot_Hurtbox.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4ad0c225fa1f77b4f8616f73085e5318
+guid: f74938f47f70d6d48867b68884a50c8c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
