+++ Assets/Scripts/Bot Parts/GunPart.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b8148ff9f8ab7cd49b4f7458c6269177
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
