+++ Assets/Scripts/Bot Parts/SwordPart.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e63390784a39344429ad79f1f1265a5f
+guid: c417a01da61a7334da9e153f3e844f28
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
