+++ Assets/Scripts/Bot Parts/SwordPart/SwordPart.cs.meta
