--- Assets/Scripts/Collision/Base_Hitbox.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b8834875cdbc49e4b9e770eb6e2ce75f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
