--- Assets/Scripts/Collision/Base_Hitresponder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a2c49d2290590df4ab5c077ece096970
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
