--- Assets/Scripts/Collision/Base_Hurtbox.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a20d768e08c36f8439d842256e29db05
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
