--- Assets/Scripts/Collision/Base_Hurtresponder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 40f6d4fc2ceee704a90b2af4bbaa058f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
