--- Assets/Scripts/Collision/Testing BotPart/Bullet/DEMO.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ce14fb4ed138e264fa9dae741cca9209
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
