--- Assets/Scripts/Collision/Testing BotPart/Bullet/Hitresponder_Bullet.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 13959f97a7a45254c81316a6f3ad1727
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
