--- Assets/Scripts/Collision/Testing BotPart/Enemy Bot/Enemy_Hurtresponder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bad668ad5c6fde943bd36af4cc2f70a1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
