--- Assets/Scripts/Collision/Testing BotPart/HitboxSword.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1eb208b7e3766674da6445acc030321e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
