--- Assets/Scripts/Collision/Testing BotPart/Sword/Sword_Hitresponder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f3565b76b46917b4a97334fbf2128249
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
