+++ Assets/Scripts/MenuScripts/BotCustomizer.cs
@@ -1,6 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using System.Linq; // List.Any()
 
 public class BotCustomizer : MonoBehaviour
 {
@@ -9,14 +10,14 @@ public class BotCustomizer : MonoBehaviour
     public GameObject targetSlot;
     //Used to set the slotposition of this instance
     [SerializeField] SlotPosition slotPosition;
-    public List<GameObject> options = new List<GameObject>();
+    public List<GameObject> options = default(List<GameObject>);
     //used to get the slots that belong to the bot
     Slots slots = default(Slots);
     private int currentOption = 0;
 
     private void Awake()
     {
-        slots =  targetBot.GetComponent<BotController>().slots;
+        slots = targetBot.GetComponent<BotController>().slots;
     }
     public void NextOption() {
         currentOption++;
@@ -35,6 +36,7 @@ public class BotCustomizer : MonoBehaviour
     }
     private void Start()
     {
+
         targetBot = GameObject.FindGameObjectWithTag("Bot");
         slots = targetBot.GetComponent<BotController>().slots;
     }
@@ -48,9 +50,9 @@ public class BotCustomizer : MonoBehaviour
         //targetSlot = newTarget;
 
 
-
-        //Set the current option botpart to the correct slot
-        slots.SetSlotBotPart(slotPosition, options[currentOption]);
+        if ((options != null) && (options.Any()))
+                slots.SetSlotBotPart(slotPosition, options[currentOption]); //Set the current option botpart to the correct slot
+        
         
       
     }
